// src/app/models/Company/company.ts
export class Company {
    companyId?: number;
    companyName?: string;
    companyAlias?: string;
    address?: string;
    phone?: string;
    fax?: string;
    email?: string;
    website?: string;
    companyRegisterNo?: string;
    isActive?: number;
    createBy?: number;
    createDate?: Date;
    updateBy?: number;
    updateDate?: Date;
}